[P2P-000] 메인으로 머지

+++ src/main/java/com/prgrms/p2p/domain/course/dto/DetailCourseResponse.java
@@ -30,7 +30,7 @@ public class DetailCourseResponse {
   @Builder.Default
   private List<CoursePlaceResponse> places = new ArrayList<>();
   private Integer likes;
-  private Integer comments;
+  private Long comments;
   private Boolean isLiked;
   private Boolean isBookmarked;
   private Long userId;
